Fixed duplicate HTML element IDs of folded elements in sidebar and wikipage.
See #47. Folded IDs are now generated in a centralized helper plugin. Also
the ID now includes the page name. As the page name is empty when the sidebar
is generated, IDs generated for the sidebar differ from those which are
generated for the wikipage. The JavaScript code for 'fold/unfold all' had to
be adjusted and now works on the classes only - independent from the element ID.